Workspace trust refactoring (#116354)
* Moved service to workbench layer

* Remove reset method and fix command/menu names

* Moved menu group
