+++ src/vs/workbench/contrib/update/browser/update.ts
@@ -434,7 +434,7 @@ export class UpdateContribution extends Disposable implements IWorkbenchContribu
 	private registerGlobalActivityActions(): void {
 		CommandsRegistry.registerCommand('update.check', () => this.updateService.checkForUpdates(this.environmentService.sessionId));
 		MenuRegistry.appendMenuItem(MenuId.GlobalActivity, {
-			group: '6_update',
+			group: '7_update',
 			command: {
 				id: 'update.check',
 				title: nls.localize('checkForUpdates', "Check for Updates...")
@@ -444,7 +444,7 @@ export class UpdateContribution extends Disposable implements IWorkbenchContribu
 
 		CommandsRegistry.registerCommand('update.checking', () => { });
 		MenuRegistry.appendMenuItem(MenuId.GlobalActivity, {
-			group: '6_update',
+			group: '7_update',
 			command: {
 				id: 'update.checking',
 				title: nls.localize('checkingForUpdates', "Checking for Updates..."),
@@ -455,7 +455,7 @@ export class UpdateContribution extends Disposable implements IWorkbenchContribu
 
 		CommandsRegistry.registerCommand('update.downloadNow', () => this.updateService.downloadUpdate());
 		MenuRegistry.appendMenuItem(MenuId.GlobalActivity, {
-			group: '6_update',
+			group: '7_update',
 			command: {
 				id: 'update.downloadNow',
 				title: nls.localize('download update_1', "Download Update (1)")
@@ -465,7 +465,7 @@ export class UpdateContribution extends Disposable implements IWorkbenchContribu
 
 		CommandsRegistry.registerCommand('update.downloading', () => { });
 		MenuRegistry.appendMenuItem(MenuId.GlobalActivity, {
-			group: '6_update',
+			group: '7_update',
 			command: {
 				id: 'update.downloading',
 				title: nls.localize('DownloadingUpdate', "Downloading Update..."),
@@ -476,7 +476,7 @@ export class UpdateContribution extends Disposable implements IWorkbenchContribu
 
 		CommandsRegistry.registerCommand('update.install', () => this.updateService.applyUpdate());
 		MenuRegistry.appendMenuItem(MenuId.GlobalActivity, {
-			group: '6_update',
+			group: '7_update',
 			command: {
 				id: 'update.install',
 				title: nls.localize('installUpdate...', "Install Update... (1)")
@@ -486,7 +486,7 @@ export class UpdateContribution extends Disposable implements IWorkbenchContribu
 
 		CommandsRegistry.registerCommand('update.updating', () => { });
 		MenuRegistry.appendMenuItem(MenuId.GlobalActivity, {
-			group: '6_update',
+			group: '7_update',
 			command: {
 				id: 'update.updating',
 				title: nls.localize('installingUpdate', "Installing Update..."),
@@ -497,7 +497,7 @@ export class UpdateContribution extends Disposable implements IWorkbenchContribu
 
 		CommandsRegistry.registerCommand('update.restart', () => this.updateService.quitAndInstall());
 		MenuRegistry.appendMenuItem(MenuId.GlobalActivity, {
-			group: '6_update',
+			group: '7_update',
 			command: {
 				id: 'update.restart',
 				title: nls.localize('restartToUpdate', "Restart to Update (1)")
@@ -541,7 +541,7 @@ export class SwitchProductQualityContribution extends Disposable implements IWor
 				}
 			});
 			MenuRegistry.appendMenuItem(MenuId.GlobalActivity, {
-				group: '6_update',
+				group: '7_update',
 				command: {
 					id: commandId,
 					title: newQuality === 'insider' ? nls.localize('switchToInsiders', "Switch to Insiders Version...") : nls.localize('switchToStable', "Switch to Stable Version...")
